requirements.txt
django==1.10.3
django-brake==1.5.2
djangorestframework==3.6.3
jsonfield==2.0.2
